examples: add a trailing slash for tftp proxy http base url
When compiling with golang-1.7, a tftp request returns the below error.

    ERR: http request returned status 400 BAD REQUEST

As explained in https://github.com/bwalex/tftp-http-proxy/issues/2, we
need to avoid the 301 HTTP redirect by either using the correct url(with
a trailing slash) to http-base-url or compiling with golang >=1.8.

Using a trailing slash seems to be more generic solution. I validated
that It works with both golang-1.7 and golang-1.8.

Signed-off-by: Chase Qi <chase.qi@linaro.org>